消费后台，生成验证码图片后台

+++ src/main/java/com/baremind/Cards.java
@@ -99,7 +99,7 @@ public class Cards {
                         exist.setUserId(userId);
                     }
 
-                    Double amount = card.getAmount();
+                    Integer amount = card.getAmount();
                     if (amount != null) {
                         exist.setAmount(amount);
                     }
